+++ clang/lib/AST/ExprConstant.cpp
@@ -59,6 +59,7 @@
 #include "llvm/ADT/Sequence.h"
 #include "llvm/ADT/SmallBitVector.h"
 #include "llvm/ADT/StringExtras.h"
+#include "llvm/ADT/bit.h"
 #include "llvm/Support/Casting.h"
 #include "llvm/Support/Debug.h"
 #include "llvm/Support/SaveAndRestore.h"
@@ -766,12 +767,14 @@ namespace {
 
     /// Get the kind of the allocation. This must match between allocation
     /// and deallocation.
-    Kind getKind() const {
+    static Kind kindOfExpr(const Expr *AllocExpr) {
       if (auto *NE = dyn_cast<CXXNewExpr>(AllocExpr))
         return NE->isArray() ? ArrayNew : New;
       assert(isa<CallExpr>(AllocExpr));
       return StdAllocator;
     }
+
+    Kind getKind() const { return kindOfExpr(AllocExpr); }
   };
 
   struct DynAllocOrder {
@@ -1956,13 +1959,92 @@ APValue &CallStackFrame::createLocal(APValue::LValueBase Base, const void *Key,
   return Result;
 }
 
+uint64_t GetAlignOfDynamicAlloc(const ASTContext &Ctx, QualType AllocType,
+                                const Expr *AllocExpr) {
+  const TargetInfo &TI = Ctx.getTargetInfo();
+  const uint64_t DefaultNewAlign = TI.getNewAlign();
+  const uint64_t MaxFundamentalAlign =
+      std::max(TI.getLongLongAlign(), TI.getLongDoubleAlign());
+  const DynAlloc::Kind AllocKind = DynAlloc::kindOfExpr(AllocExpr);
+
+  const uint64_t AllocSize = Ctx.getTypeSize(AllocType);
+
+  // Allocating a zero-sized array is allowed, however it doesn't have
+  // any alignment guarantees.
+  if ((AllocKind == DynAlloc::Kind::ArrayNew ||
+       AllocKind == DynAlloc::Kind::StdAllocator) &&
+      AllocSize == 0)
+    return TI.getCharWidth();
+
+  // For the non-array form, the size of the allocated object should be
+  // positive.
+  assert(AllocSize > 0 && "Unknown size for allocated type!");
+
+  const uint64_t TypeAlignment = Ctx.getTypeAlign(AllocType);
+  assert(TypeAlignment > 0 && "Unknown alignment for allocated type!");
+
+  // For new-extended alignment the ::operator new overload with
+  // std::align_val_t parameter is used. According to C++
+  // [basic.stc.dynamic.allocation]p3.1 this overload returns memory
+  // according to the requested alignment. No stricter guarantees are
+  // made.
+  if (TypeAlignment > DefaultNewAlign)
+    return TypeAlignment;
+
+  switch (AllocKind) {
+  // According to C++ [allocator.members]p5 it is unspecified how the
+  // memory obtained from ::operator new is used by
+  // std::allocator::allocate, therefore be conservative here.
+  case DynAlloc::Kind::StdAllocator:
+    return TypeAlignment;
+
+  // The non-array form of new does not permit allocation overhead and
+  // therefore provides alignment as guaranteed by ::operator new.
+  // According to C++ [basic.stc.dynamic.allocation]p3.3 the allocation
+  // is suitably aligned for all objects without new-extended alignment
+  // with the exact size of the allocation. An object of the exact size
+  // AllocSize can have alignment of at most the lowest bit set in
+  // AllocSize.
+  case DynAlloc::Kind::New:
+    return std::min(DefaultNewAlign,
+                    uint64_t{1} << llvm::countr_zero(AllocSize));
+
+  case DynAlloc::Kind::ArrayNew: {
+    const Type *ET = AllocType.getTypePtr()
+                         ->getArrayElementTypeNoTypeQual()
+                         ->getCanonicalTypeUnqualified()
+                         .getTypePtr();
+    // According to C++ [expr.new]p17, unless the element type of an
+    // array new expression is char, unsigned char or std::byte, the
+    // allocation may be offset into the allocation returned by
+    // ::operator new[]. Therefore no stricter alignment than the type's
+    // alignment is guaranteed. For char, unsigned char and std::byte
+    if (!ET->isSpecificBuiltinType(BuiltinType::UChar) &&
+        !ET->isSpecificBuiltinType(BuiltinType::Char_U) &&
+        !ET->isSpecificBuiltinType(BuiltinType::Char_S) && !ET->isStdByteType())
+      return TypeAlignment;
+
+    // Otherwise, the allocation is offset from the result of ::operator
+    // new[] by a multiple of the strictest fundamental alignment.
+    // According C++ [basic.stc.dynamic.allocation]p3.2 the allocation
+    // returned by ::operator new[] is suitably aligned for all objects
+    // without new-extended alignment and size up to the allocated size.
+    return std::min(
+        {DefaultNewAlign, MaxFundamentalAlign, llvm::bit_floor(AllocSize)});
+  }
+
+  default:
+    llvm_unreachable("Unhandled DynAlloc::Kind");
+  }
+}
+
 APValue *EvalInfo::createHeapAlloc(const Expr *E, QualType T, LValue &LV) {
   if (NumHeapAllocs > DynamicAllocLValue::getMaxIndex()) {
     FFDiag(E, diag::note_constexpr_heap_alloc_limit_exceeded);
     return nullptr;
   }
 
-  DynamicAllocLValue DA(NumHeapAllocs++);
+  DynamicAllocLValue DA(NumHeapAllocs++, GetAlignOfDynamicAlloc(Ctx, T, E));
   LV.set(APValue::LValueBase::getDynamicAlloc(DA, T));
   auto Result = HeapAllocs.emplace(std::piecewise_construct,
                                    std::forward_as_tuple(DA), std::tuple<>());
@@ -10341,6 +10423,8 @@ static CharUnits getBaseAlignment(EvalInfo &Info, const LValue &Value) {
     return Info.Ctx.getDeclAlign(VD);
   if (const auto *E = Value.Base.dyn_cast<const Expr *>())
     return GetAlignOfExpr(Info.Ctx, E, UETT_AlignOf);
+  if (const auto &DA = Value.Base.dyn_cast<DynamicAllocLValue>())
+    return Info.Ctx.toCharUnitsFromBits(DA.getAlign());
   return GetAlignOfType(Info.Ctx, Value.Base.getTypeInfoType(), UETT_AlignOf);
 }